Fix examples charts/dashboards and refactor
